Specify action when launching SessionDetailActivity from notification
This solves the following two problems:
1. Because we didn't provide requestCode, and the intents differ only in extras,
the system can't tell the differences between consecutive intents, which causes
different notifications to open the same session.

2. If the app task had been removed from recents, and SessionDetailActivity was
launched from a notification, after user pressed the up button in the activity
the app task would look like this:
```
Task{
  intent=Intent{
    class=SessionDetailActivity,
    action=null,
  },
  activityStack=[MainActivity],
}
```
If user then opened another notification, the system would think there's
already a task with the same intent, and thus only bring the task to the front.